Crate bon_macros

bon-macros

This is a proc-macro crate that is supposed to be a private implementation detail of the bon crate. Don’t add it to your dependencies directly! The API surface of this crate is unstable, and your code may break if you use items from bon-macros bypassing the bon crate. Instead, use the proc macros from here via the reexports in the bon crate.

Macros

map

Creates any map-like collection that implements FromIterator<(K, V)>.

set

Creates any set-like collection that implements FromIterator<T>.

Attribute Macros

bon

Companion macro for builder. You should place it on top of the impl block where you want to define methods with the builder macro.

builder

Generates a builder for the function or method it’s placed on.

Derive Macros

Builder

Derives a builder for the struct it’s placed on.